requirements.txt
unicodedata2
google-search-results
requests
beautifulsoup4
python-dotenv